restructure `Config.do_update()`
This commit organizes the `Config.do_update()` method in 3 parts:
1. Migrate old values
2. Copy values to base config
3. Validate base config

Migration could be coupled with copying values to the base config by
directly migrating to base config and only copying non-migrated values,
but it's cleaner if the steps are decoupled and the migration happens
within the source config IMO.

Validation happens afterwards on the base config, as the base config
also contains default values in case of an unspecified value in the
source config. This way, validation won't fail due to unspecified
default values.